fix(tests): client_mocks moved inside the package, imports for tests adjusted
