+++ tests/unit_tests/test_abort_button.py
@@ -2,9 +2,8 @@
 
 import pytest
 
+from bec_widgets.tests.client_mocks import mocked_client
 from bec_widgets.widgets.control.buttons.button_abort.button_abort import AbortButton
-
-from .client_mocks import mocked_client
 
 
 @pytest.fixture
