+++ tests/unit_tests/test_device_input_widgets.py
@@ -1,13 +1,12 @@
 import pytest
 from bec_lib.device import ReadoutPriority
 
+from bec_widgets.tests.client_mocks import mocked_client
 from bec_widgets.widgets.control.device_input.base_classes.device_input_base import BECDeviceFilter
 from bec_widgets.widgets.control.device_input.device_combobox.device_combobox import DeviceComboBox
 from bec_widgets.widgets.control.device_input.device_line_edit.device_line_edit import (
     DeviceLineEdit,
 )
-
-from .client_mocks import mocked_client
 
 
 @pytest.fixture
