+++ tests/unit_tests/test_filter_io.py
@@ -1,12 +1,12 @@
 import pytest
 
+from bec_widgets.tests.client_mocks import mocked_client
 from bec_widgets.utils.filter_io import FilterIO
 from bec_widgets.widgets.control.device_input.device_line_edit.device_line_edit import (
     DeviceLineEdit,
 )
 from bec_widgets.widgets.dap.dap_combo_box.dap_combo_box import DapComboBox
 
-from .client_mocks import mocked_client
 from .conftest import create_widget
