+++ tests/unit_tests/test_image_rois.py
@@ -5,6 +5,7 @@ from typing import Literal
 import numpy as np
 import pytest
 
+from bec_widgets.tests.client_mocks import mocked_client
 from bec_widgets.widgets.plots.image.image import Image
 from bec_widgets.widgets.plots.roi.image_roi import (
     CircularROI,
@@ -12,7 +13,6 @@ from bec_widgets.widgets.plots.roi.image_roi import (
     RectangularROI,
     ROIController,
 )
-from tests.unit_tests.client_mocks import mocked_client
 from tests.unit_tests.conftest import create_widget
