+++ tests/unit_tests/test_image_view_next_gen.py
@@ -3,8 +3,8 @@ import pyqtgraph as pg
 import pytest
 from qtpy.QtCore import QPointF
 
+from bec_widgets.tests.client_mocks import mocked_client
 from bec_widgets.widgets.plots.image.image import Image
-from tests.unit_tests.client_mocks import mocked_client
 from tests.unit_tests.conftest import create_widget
 
 ##################################################
