+++ tests/unit_tests/test_motor_map_next_gen.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pyqtgraph as pg
 
+from bec_widgets.tests.client_mocks import mocked_client
 from bec_widgets.widgets.plots.motor_map.motor_map import MotorMap
-from tests.unit_tests.client_mocks import mocked_client
 
 from .conftest import create_widget
