+++ tests/unit_tests/test_multi_waveform_next_gen.py
@@ -1,7 +1,7 @@
 import numpy as np
 
+from bec_widgets.tests.client_mocks import mocked_client
 from bec_widgets.widgets.plots.multi_waveform.multi_waveform import MultiWaveform
-from tests.unit_tests.client_mocks import mocked_client
 
 from .conftest import create_widget
