+++ tests/unit_tests/test_positioner_box_2d.py
@@ -2,9 +2,9 @@ from unittest import mock
 
 import pytest
 
+from bec_widgets.tests.client_mocks import mocked_client
 from bec_widgets.widgets.control.device_control.positioner_box import PositionerBox2D
 
-from .client_mocks import mocked_client
 from .conftest import create_widget
