+++ tests/unit_tests/test_reset_button.py
@@ -5,9 +5,8 @@ from unittest.mock import patch
 import pytest
 from qtpy.QtWidgets import QMessageBox
 
+from bec_widgets.tests.client_mocks import mocked_client
 from bec_widgets.widgets.control.buttons.button_reset.button_reset import ResetButton
-
-from .client_mocks import mocked_client
 
 
 @pytest.fixture
