+++ tests/unit_tests/test_stop_button.py
@@ -2,9 +2,8 @@
 
 import pytest
 
+from bec_widgets.tests.client_mocks import mocked_client
 from bec_widgets.widgets.control.buttons.stop_button.stop_button import StopButton
-
-from .client_mocks import mocked_client
 
 
 @pytest.fixture
