+++ tests/unit_tests/test_text_box_widget.py
@@ -1,8 +1,7 @@
 import pytest
 
+from bec_widgets.tests.client_mocks import mocked_client
 from bec_widgets.widgets.editors.text_box.text_box import DEFAULT_TEXT, TextBox
-
-from .client_mocks import mocked_client
 
 
 @pytest.fixture
