+++ tests/unit_tests/test_vscode_widget.py
@@ -5,9 +5,8 @@ from unittest import mock
 
 import pytest
 
+from bec_widgets.tests.client_mocks import mocked_client
 from bec_widgets.widgets.editors.vscode.vscode import VSCodeEditor
-
-from .client_mocks import mocked_client
 
 
 @pytest.fixture
