+++ tests/unit_tests/test_waveform.py
@@ -12,19 +12,19 @@ from pyqtgraph.graphicsItems.DateAxisItem import DateAxisItem
 from qtpy.QtCore import QTimer
 from qtpy.QtWidgets import QApplication, QCheckBox, QDialog, QDialogButtonBox, QDoubleSpinBox
 
-from bec_widgets.widgets.plots.plot_base import UIMode
-from bec_widgets.widgets.plots.waveform.curve import DeviceSignal
-from bec_widgets.widgets.plots.waveform.waveform import Waveform
-from bec_widgets.widgets.services.scan_history_browser.scan_history_browser import (
-    ScanHistoryBrowser,
-)
-from tests.unit_tests.client_mocks import (
+from bec_widgets.tests.client_mocks import (
     DummyData,
     create_dummy_scan_item,
     dap_plugin_message,
     inject_scan_history,
     mocked_client,
     mocked_client_with_dap,
+)
+from bec_widgets.widgets.plots.plot_base import UIMode
+from bec_widgets.widgets.plots.waveform.curve import DeviceSignal
+from bec_widgets.widgets.plots.waveform.waveform import Waveform
+from bec_widgets.widgets.services.scan_history_browser.scan_history_browser import (
+    ScanHistoryBrowser,
 )
 
 from .conftest import create_widget
