+++ tests/unit_tests/test_website_widget.py
@@ -1,9 +1,8 @@
 import pytest
 from qtpy.QtCore import QUrl
 
+from bec_widgets.tests.client_mocks import mocked_client
 from bec_widgets.widgets.editors.website.website import WebsiteWidget
-
-from .client_mocks import mocked_client
 
 
 @pytest.fixture
